# frontend/src/server/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tenders API",
    "description": "Tender-response backend. Service-authenticated: the Next.js server sets X-Actor and X-Org-Id on every request and the browser never calls this API directly. Without a proxy (tests, the harness, curl) X-Org-Id defaults to the seeded organisation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/organisations": {
      "post": {
        "tags": [
          "organisations"
        ],
        "summary": "Provision",
        "description": "Caller-generated IDs make retries safe across the two databases.",
        "operationId": "provision_organisations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/current": {
      "get": {
        "tags": [
          "organisations"
        ],
        "summary": "Current",
        "description": "The organisation X-Org-Id names, so the front end can label a synthetic business.",
        "operationId": "current_organisations_current_get",
        "parameters": [
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tenders/{tender_id}/requirements": {
      "get": {
        "tags": [
          "requirements"
        ],
        "summary": "List Requirements",
        "description": "{summary: {green, amber, red, unrated, total, suggested_unrated}, job: the latest\nextract_requirements job id or null, requirements: [...]} in document order.",
        "operationId": "list_requirements_tenders__tender_id__requirements_get",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tender Id"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenders/{tender_id}/requirements/rescan": {
      "post": {
        "tags": [
          "requirements"
        ],
        "summary": "Rescan Requirements",
        "description": "Enqueue an extract_requirements run over every ready tender document, or return the one\nalready queued. Ratings, comments and owners are kept by the re-run.",
        "operationId": "rescan_requirements_tenders__tender_id__requirements_rescan_post",
        "parameters": [
          {
            "name": "tender_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tender Id"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RescanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/requirements/{requirement_id}": {
      "patch": {
        "tags": [
          "requirements"
        ],
        "summary": "Patch Requirement",
        "description": "Set any of compliance_class, compliant_by, comment and owner. A change of\ncompliance_class stamps rated_by and rated_at from X-Actor and writes a\n``requirement_rated`` event with from and to; clearing it clears both stamps. Moving to A\nor C clears compliant_by unless the same request sets it.",
        "operationId": "patch_requirement_requirements__requirement_id__patch",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Requirement Id"
            }
          },
          {
            "name": "X-Org-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganisationCreate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "synthetic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthetic"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "OrganisationCreate"
      },
      "OrganisationRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "synthetic"
        ],
        "title": "OrganisationRecord"
      },
...
      "RequirementList": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/RequirementSummary"
          },
          "job": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementRecord"
            },
            "type": "array",
            "title": "Requirements"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "RequirementList"
      },
      "RequirementLocator": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "section_id": {
            "type": "string",
            "format": "uuid",
            "title": "Section Id"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "table": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table"
          },
          "cell_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell Ref"
          },
          "heading_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Heading Path"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "section_id"
        ],
        "title": "RequirementLocator",
        "description": "Where the requirement's text sits in its tender document; ``start`` and ``end`` are null\nwhen the text was not located in the section."
      },
      "RequirementPatch": {
        "properties": {
          "compliance_class": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Class"
          },
          "compliant_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliant By"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          }
        },
        "type": "object",
        "title": "RequirementPatch",
        "description": "Any of the human fields. ``compliance_class: null`` clears the rating."
      },
      "RequirementRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tender_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tender Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "document_filename": {
            "type": "string",
            "title": "Document Filename"
          },
          "tender_doc_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tender Doc Kind"
          },
          "locator": {
            "$ref": "#/components/schemas/RequirementLocator"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "must",
                  "should",
                  "could"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics"
          },
          "suggested_class": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Class"
          },
          "suggestion": {
            "additionalProperties": true,
            "type": "object",
            "title": "Suggestion"
          },
          "compliance_class": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Class"
          },
          "compliant_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliant By"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "rated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rated By"
          },
          "rated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rated At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tender_id",
          "document_id",
          "document_filename",
          "locator",
          "text",
          "order_index",
          "created_at",
          "updated_at"
        ],
        "title": "RequirementRecord"
      },
      "RequirementSummary": {
        "properties": {
          "green": {
            "type": "integer",
            "title": "Green",
            "default": 0
          },
          "amber": {
            "type": "integer",
            "title": "Amber",
            "default": 0
          },
          "red": {
            "type": "integer",
            "title": "Red",
            "default": 0
          },
          "unrated": {
            "type": "integer",
            "title": "Unrated",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "suggested_unrated": {
            "type": "integer",
            "title": "Suggested Unrated",
            "default": 0
          }
        },
        "type": "object",
        "title": "RequirementSummary",
        "description": "``suggested_unrated``: unrated requirements that carry an AI suggestion."
      },
      "RescanResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "RescanResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}